+++ library/src/actions/title/title.test-d.ts
@@ -1,0 +1,25 @@
+import { describe, expectTypeOf, test } from 'vitest';
+import type { InferInput, InferIssue, InferOutput } from '../../types/index.ts';
+import { title, type TitleAction } from './title.ts';
+
+describe('title', () => {
+  type Action = TitleAction<string, 'text'>;
+
+  test('should return action object', () => {
+    expectTypeOf(title<string, 'text'>('text')).toEqualTypeOf<Action>();
+  });
+
+  describe('should infer correct types', () => {
+    test('of input', () => {
+      expectTypeOf<InferInput<Action>>().toEqualTypeOf<string>();
+    });
+
+    test('of output', () => {
+      expectTypeOf<InferOutput<Action>>().toEqualTypeOf<string>();
+    });
+
+    test('of issue', () => {
+      expectTypeOf<InferIssue<Action>>().toEqualTypeOf<never>();
+    });
+  });
+});