+++ library/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Add `metadata` action to add custom metadata to a schema
+- Add `title` metadata action to add a title to a schema (discussion #826)
 - Add `decimal` action to validate integer and float strings (pull request #823)
 - Rename `NoPipe` type to `SchemaWithoutPipe`
 - Rename `decimal` action to `digits` (pull request #823)